[refactor] App.js 다이어트 + NotionTitle.js 이벤트 델리게이션 분해
1. App에서 findDocumentDataById 추출

2. 이벤트는 DOM 내에서만 적용되도록 분해
